pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "books-we-love"
version = "0.1.0"
description = "CLI tools for downloading NPR 'Books We Love' JSON data and tracking external book/author lookups"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0,<3.0.0",
    "readarr-py==1.1.0",
    "jsonpath-ng>=1.6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
books-we-love = "books_we_love.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["books_we_love*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

